Move MemorySegmentInputStream and make it public
